Init define changes.
