+++ AMBuilder
@@ -8,6 +8,8 @@ project = builder.LibraryProject(projectName)
 project.sources += [
   os.path.join(Extension.sm_root, "public", "smsdk_ext.cpp"),
   'src/extension.cpp',
+  'src/utils.cpp',
+  'src/wrappers.cpp',
   'src/SI/boomer.cpp',
   'src/SI/smoker.cpp'
 ]
